generative-ai/24.LangChain2/Project3/requirements.txt
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.6
openai>=1.30,<2.0
tiktoken>=0.9,<1.0
pymongo==4.17.0
python-dotenv==1.0.0
pypdf==4.2.0
colorama==0.4.6
